+++ src/plugins/postcss-icss-parser.js
@@ -1,9 +1,8 @@
 import postcss from 'postcss';
 import { extractICSS, replaceValueSymbols, replaceSymbols } from 'icss-utils';
 import loaderUtils from 'loader-utils';
-import cc from 'camelcase';
 
-import { getImportItemCode } from '../utils';
+import { getExportItemCode, getImportItemCode } from '../utils';
 
 const pluginName = 'postcss-icss-parser';
 
@@ -15,56 +14,6 @@ function hasImportMessage(messages, url) {
       message.item.url === url &&
       message.item.media === ''
   );
-}
-
-function camelCase(str) {
-  return cc(str);
-}
-
-function dashesCamelCase(str) {
-  return str.replace(/-+(\w)/g, (match, firstLetter) =>
-    firstLetter.toUpperCase()
-  );
-}
-
-function getExportItem(key, value, localsStyle) {
-  let targetKey;
-  const items = [];
-
-  function addEntry(k) {
-    items.push(`\t${JSON.stringify(k)}: ${JSON.stringify(value)}`);
-  }
-
-  switch (localsStyle) {
-    case 'camelCase':
-      addEntry(key);
-      targetKey = camelCase(key);
-
-      if (targetKey !== key) {
-        addEntry(targetKey);
-      }
-      break;
-    case 'camelCaseOnly':
-      addEntry(camelCase(key));
-      break;
-    case 'dashes':
-      addEntry(key);
-      targetKey = dashesCamelCase(key);
-
-      if (targetKey !== key) {
-        addEntry(targetKey);
-      }
-      break;
-    case 'dashesOnly':
-      addEntry(dashesCamelCase(key));
-      break;
-    case 'asIs':
-    default:
-      addEntry(key);
-      break;
-  }
-
-  return items;
 }
 
 export default postcss.plugin(
@@ -118,9 +67,7 @@ export default postcss.plugin(
 
         result.messages.push({
           pluginName,
-          export: getExportItem(name, value, options.exportLocalsStyle).join(
-            ',\n'
-          ),
+          export: getExportItemCode(name, value, options.exportLocalsStyle),
           type: 'export',
           item: { name, value },
         });
